requirements.txt
# === MINIMAL REQUIREMENTS FOR TRADING BOTS ===
# Essential dependencies only

python-binance==1.0.19
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
cryptography